[DRAFT]: Divided AA into drought and storm layers (#1370)
* Divided AA into drought and storm layers and panels

* fix misspelling and try to fix url update AA selection change

* updated switch AA layers in url

* Fix tests

* reset start date on switch AA action

* do not show data in the timeline when layer anticipatory_action_storm is selected

* isolate AA drought specific legen items, setup empty legend for AA storm

* clean AAstormPanel

* move AA drought specific components to AADroughtPanel folder

* do not show elements on the map for storm AA

* fix linting issues

* fix tests

* enable the display of 2 aa layers only for aa_drought

* add json data in a temporary file

---------
