+++ frontend/src/components/MapView/Layers/AnticipatoryActionStormLayer/index.tsx
@@ -1,0 +1,14 @@
+import React from 'react';
+import { AnticipatoryActionLayerProps } from 'config/types';
+import { useDefaultDate } from 'utils/useDefaultDate';
+
+const AnticipatoryActionStormLayer = React.memo(({ layer }: LayersProps) => {
+  useDefaultDate(layer.id);
+
+  return null;
+});
+export interface LayersProps {
+  layer: AnticipatoryActionLayerProps;
+}
+
+export default AnticipatoryActionStormLayer;
