+++ frontend/src/components/MapView/LeftPanel/AnticipatoryActionPanel/AnticipatoryActionDroughtPanel/DistrictView/index.test.tsx
@@ -3,9 +3,9 @@ import { Provider } from 'react-redux';
 import { BrowserRouter } from 'react-router-dom';
 import configureStore from 'redux-mock-store';
 import { AnticipatoryActionState } from 'context/anticipatoryActionStateSlice/types';
-import { Panel } from 'context/leftPanelStateSlice';
+import { Panel } from 'config/types';
 import DistrictView from '.';
-import { defaultDialogs, mockAAData } from '../test.utils';
+import { defaultDialogs, mockAAData } from '../../test.utils';
 import { districtViewTransform } from './utils';
 
 const mockStore = configureStore([]);
@@ -42,7 +42,7 @@ const store = mockStore({
     data: mockAAData,
   },
   leftPanelState: {
-    tabValue: Panel.AnticipatoryAction,
+    tabValue: Panel.AnticipatoryActionDrought,
   },
 });
