+++ frontend/src/components/MapView/LeftPanel/AnticipatoryActionPanel/AnticipatoryActionDroughtPanel/Forecast/index.test.tsx
@@ -3,8 +3,8 @@ import { Provider } from 'react-redux';
 import { BrowserRouter } from 'react-router-dom';
 import configureStore from 'redux-mock-store';
 import { AnticipatoryActionState } from 'context/anticipatoryActionStateSlice/types';
-import { Panel } from 'context/leftPanelStateSlice';
-import { defaultDialogs, mockAAData } from '../test.utils';
+import { Panel } from 'config/types';
+import { defaultDialogs, mockAAData } from '../../test.utils';
 import { forecastTransform } from './utils';
 import Forecast from '.';
 
@@ -42,7 +42,7 @@ const store = mockStore({
     data: mockAAData,
   },
   leftPanelState: {
-    tabValue: Panel.AnticipatoryAction,
+    tabValue: Panel.AnticipatoryActionDrought,
   },
 });
