+++ frontend/src/components/MapView/LeftPanel/AnticipatoryActionPanel/AnticipatoryActionDroughtPanel/HomeTable/index.test.tsx
@@ -2,9 +2,9 @@ import { render } from '@testing-library/react';
 import { Provider } from 'react-redux';
 import { BrowserRouter } from 'react-router-dom';
 import configureStore from 'redux-mock-store';
-import { Panel } from 'context/leftPanelStateSlice';
+import { Panel } from 'config/types';
 import HomeTable from '.';
-import { defaultDialogs, mockAARenderedDistricts } from '../test.utils';
+import { defaultDialogs, mockAARenderedDistricts } from '../../test.utils';
 
 const mockStore = configureStore([]);
 
@@ -33,7 +33,7 @@ const store = mockStore({
     windowRanges: { 'Window 2': { start: '2023-08-01', end: '2023-12-01' } },
   },
   leftPanelState: {
-    tabValue: Panel.AnticipatoryAction,
+    tabValue: Panel.AnticipatoryActionDrought,
   },
 });
