+++ frontend/src/components/MapView/LeftPanel/AnticipatoryActionPanel/AnticipatoryActionDroughtPanel/Timeline/index.test.tsx
@@ -3,8 +3,8 @@ import { Provider } from 'react-redux';
 import { BrowserRouter } from 'react-router-dom';
 import configureStore from 'redux-mock-store';
 import { AnticipatoryActionState } from 'context/anticipatoryActionStateSlice/types';
-import { Panel } from 'context/leftPanelStateSlice';
-import { defaultDialogs, mockAAData } from '../test.utils';
+import { Panel } from 'config/types';
+import { defaultDialogs, mockAAData } from '../../test.utils';
 import { timelineTransform } from './utils';
 import Timeline from '.';
 
@@ -42,7 +42,7 @@ const store = mockStore({
     data: mockAAData,
   },
   leftPanelState: {
-    tabValue: Panel.AnticipatoryAction,
+    tabValue: Panel.AnticipatoryActionDrought,
   },
 });
