+++ frontend/src/components/MapView/LeftPanel/index.tsx
@@ -2,14 +2,19 @@ import { Drawer, Theme, createStyles, makeStyles } from '@material-ui/core';
 import React, { memo } from 'react';
 import { useDispatch, useSelector } from 'react-redux';
 import {
-  Panel,
   leftPanelTabValueSelector,
   setTabValue,
 } from 'context/leftPanelStateSlice';
 import {
-  AALayerId,
+  AnticipatoryAction,
+  AnticipatoryActionLayerProps,
+  Panel,
+} from 'config/types';
+import {
+  AALayerIds,
   LayerDefinitions,
   areChartLayersAvailable,
+  isAnticipatoryActionLayer,
 } from 'config/utils';
 import { getUrlKey, useUrlHistory } from 'utils/url-utils';
 import { layersSelector, mapSelector } from 'context/mapStateSlice/selectors';
@@ -24,10 +29,14 @@ import {
   updateLayersCapabilities,
 } from 'context/serverStateSlice';
 import { getAAAvailableDatesCombined } from 'utils/server-utils';
+import { updateDateRange } from 'context/mapStateSlice';
 import AnalysisPanel from './AnalysisPanel';
 import ChartsPanel from './ChartsPanel';
 import TablesPanel from './TablesPanel';
-import AnticipatoryActionPanel from './AnticipatoryActionPanel';
+import {
+  AnticipatoryActionDroughtPanel,
+  AnticipatoryActionStormPanel,
+} from './AnticipatoryActionPanel';
 import LayersPanel from './layersPanel';
 import { areTablesAvailable, isAnticipatoryActionAvailable } from './utils';
 import { toggleRemoveLayer } from './layersPanel/MenuItem/MenuSwitch/SwitchItem/utils';
@@ -75,19 +84,22 @@ const LeftPanel = memo(() => {
   // Sync serverAvailableDates with AAAvailableDates when the latter updates.
   React.useEffect(() => {
     if (AAAvailableDates) {
-      dispatch(
-        updateLayersCapabilities({
-          ...serverAvailableDates,
-          [AALayerId]: getAAAvailableDatesCombined(AAAvailableDates),
+      const updatedCapabilities = AALayerIds.reduce(
+        (acc, layerId) => ({
+          ...acc,
+          [layerId]: getAAAvailableDatesCombined(AAAvailableDates),
         }),
+        { ...serverAvailableDates },
       );
+      dispatch(updateLayersCapabilities(updatedCapabilities));
     }
     // To avoid an infinite loop, we only want to run this effect when AAAvailableDates changes.
     // eslint-disable-next-line react-hooks/exhaustive-deps
   }, [AAAvailableDates, dispatch]);
 
   const AALayerInUrl = React.useMemo(
-    () => selectedLayers.find(x => x.id === AALayerId),
+    () =>
+      selectedLayers.find(x => AALayerIds.includes(x.id as AnticipatoryAction)),
     [selectedLayers],
   );
 
@@ -96,18 +108,22 @@ const LeftPanel = memo(() => {
     // TODO: update  Object.keys(AAData).length === 0 condition with something more solid
     // Move to AA tab when directly linked there
     if (
-      tabValue !== Panel.AnticipatoryAction &&
+      !isAnticipatoryActionLayer(tabValue) &&
       AALayerInUrl !== undefined &&
       Object.keys(AAData['Window 1']).length === 0
     ) {
-      dispatch(setTabValue(Panel.AnticipatoryAction));
+      if (AALayerInUrl.id === AnticipatoryAction.drought) {
+        dispatch(setTabValue(Panel.AnticipatoryActionDrought));
+      } else if (AALayerInUrl.id === AnticipatoryAction.storm) {
+        dispatch(setTabValue(Panel.AnticipatoryActionStorm));
+      }
     }
   }, [AAData, AALayerInUrl, dispatch, tabValue]);
 
   // Remove from url when leaving from AA tab
   React.useEffect(() => {
     if (
-      tabValue !== Panel.AnticipatoryAction &&
+      !isAnticipatoryActionLayer(tabValue) &&
       tabValue !== Panel.None &&
       AALayerInUrl !== undefined &&
       Object.keys(AAData['Window 1']).length !== 0
@@ -124,23 +140,37 @@ const LeftPanel = memo(() => {
 
   // fetch csv data when loading AA page
   React.useEffect(() => {
-    if (tabValue !== Panel.AnticipatoryAction) {
+    if (!isAnticipatoryActionLayer(tabValue)) {
       return;
     }
     dispatch(loadAAData());
   }, [dispatch, tabValue]);
 
-  // Add layers to url
+  // Add or switch AA layers in url
   React.useEffect(() => {
-    if (tabValue !== Panel.AnticipatoryAction) {
+    if (!isAnticipatoryActionLayer(tabValue)) {
+      return;
+    }
+    const selectedLayerId = AALayerIds.find(x => x === tabValue);
+    if (!selectedLayerId) {
       return;
     }
 
-    const layer = LayerDefinitions[AALayerId];
-
-    // Add to url when getting to AA tab
-    if (AALayerInUrl !== undefined || !layer) {
+    const layer = LayerDefinitions[
+      selectedLayerId
+    ] as AnticipatoryActionLayerProps;
+    if (!layer || AALayerInUrl?.id === layer.id) {
       return;
+    }
+
+    if (AALayerInUrl) {
+      toggleRemoveLayer(
+        AALayerInUrl,
+        map,
+        getUrlKey(AALayerInUrl),
+        dispatch,
+        removeLayerFromUrl,
+      );
     }
 
     const updatedUrl = appendLayerToUrl(
@@ -148,8 +178,11 @@ const LeftPanel = memo(() => {
       selectedLayers,
       layer,
     );
-
     updateHistory(getUrlKey(layer), updatedUrl);
+
+    // Reset startDate
+    dispatch(updateDateRange({ startDate: undefined }));
+
     // url does not instantly update. updateHistory and appendLayerToUrl functions re-trigger useEffect, before selected layers is set
     // eslint-disable-next-line react-hooks/exhaustive-deps
   }, [selectedLayers, tabValue, dispatch, AALayerInUrl]);
@@ -196,11 +229,21 @@ const LeftPanel = memo(() => {
     if (!isAnticipatoryActionAvailable) {
       return null;
     }
-    return (
-      <TabPanel value={tabValue} index={Panel.AnticipatoryAction}>
-        <AnticipatoryActionPanel />
-      </TabPanel>
-    );
+    if (tabValue === Panel.AnticipatoryActionDrought) {
+      return (
+        <TabPanel value={tabValue} index={Panel.AnticipatoryActionDrought}>
+          <AnticipatoryActionDroughtPanel />
+        </TabPanel>
+      );
+    }
+    if (tabValue === Panel.AnticipatoryActionStorm) {
+      return (
+        <TabPanel value={tabValue} index={Panel.AnticipatoryActionStorm}>
+          <AnticipatoryActionStormPanel />
+        </TabPanel>
+      );
+    }
+    return null;
   }, [tabValue]);
 
   return (
