+++ frontend/src/components/MapView/Map/index.tsx
@@ -14,7 +14,7 @@ import { setMap } from 'context/mapStateSlice';
 import { appConfig } from 'config';
 import useMapOnClick from 'components/MapView/useMapOnClick';
 import { setBounds, setLocation } from 'context/mapBoundaryInfoStateSlice';
-import { DiscriminateUnion, LayerKey, LayerType } from 'config/types';
+import { DiscriminateUnion, LayerKey, LayerType, Panel } from 'config/types';
 import { setLoadingLayerIds } from 'context/mapTileLoadingStateSlice';
 import {
   firstBoundaryOnView,
@@ -24,7 +24,8 @@ import {
 import { mapSelector } from 'context/mapStateSlice/selectors';
 import {
   AdminLevelDataLayer,
-  AnticipatoryActionLayer,
+  AnticipatoryActionDroughtLayer,
+  AnticipatoryActionStormLayer,
   BoundaryLayer,
   CompositeLayer,
   ImpactLayer,
@@ -37,7 +38,7 @@ import MapGL, { MapEvent, MapRef } from 'react-map-gl/maplibre';
 import { MapSourceDataEvent, Map as MaplibreMap } from 'maplibre-gl';
 
 import 'maplibre-gl/dist/maplibre-gl.css';
-import { Panel, leftPanelTabValueSelector } from 'context/leftPanelStateSlice';
+import { leftPanelTabValueSelector } from 'context/leftPanelStateSlice';
 import { mapStyle } from './utils';
 
 type LayerComponentsMap<U extends LayerType> = {
@@ -54,8 +55,11 @@ const componentTypes: LayerComponentsMap<LayerType> = {
   point_data: { component: PointDataLayer },
   static_raster: { component: StaticRasterLayer },
   composite: { component: CompositeLayer },
-  anticipatory_action: {
-    component: AnticipatoryActionLayer,
+  anticipatory_action_drought: {
+    component: AnticipatoryActionDroughtLayer,
+  },
+  anticipatory_action_storm: {
+    component: AnticipatoryActionStormLayer,
   },
 };
 
@@ -227,7 +231,10 @@ const MapComponent = memo(() => {
         return createElement(component as any, {
           key: layer.id,
           layer,
-          before: getBeforeId(index, layer.type === 'anticipatory_action'),
+          before: getBeforeId(
+            index,
+            layer.type.startsWith('anticipatory_action'),
+          ),
         });
       })}
       <AnalysisLayer before={firstBoundaryId} />
