+++ frontend/src/components/NavBar/PrintImage/index.test.tsx
@@ -2,7 +2,7 @@ import { render } from '@testing-library/react';
 import configureStore from 'redux-mock-store';
 import { Provider } from 'react-redux';
 import { BrowserRouter } from 'react-router-dom';
-import { Panel } from 'context/leftPanelStateSlice';
+import { Panel } from 'config/types';
 import Download from '.';
 
 const mockStore = configureStore([]);
@@ -32,7 +32,7 @@ test('renders as expected', () => {
       isMapLayerActive: false,
     },
     leftPanelState: {
-      tabValue: Panel.AnticipatoryAction,
+      tabValue: Panel.AnticipatoryActionDrought,
     },
   });
